backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.23.2
packaging==24.2
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
Telethon==1.38.1
typing_extensions==4.12.2
Werkzeug==3.1.3
